Add _patches, rename _undocumented, unsort args.

+++ tasks/update.rake
@@ -11,7 +11,10 @@ namespace :api do
 
     desc 'Delete all generated files except undocumented ones.'
     task :clean_files, :dirs do |_t, args|
-      files = Dir["./{#{Array(args[:dirs]).join(',')}}/*"].grep_v(%r{/undocumented\b})
+      raise 'Missing :dirs' unless args[:dirs]
+
+      # keep _undocumented and _patches
+      files = Dir["./{#{Array(args[:dirs]).join(',')}}/*"].grep_v(%r{/_\w*\b})
       FileUtils.rm_rf files
     end
   end
